+++ app/Providers/RouteServiceProvider.php
@@ -11,25 +11,16 @@ use Illuminate\Support\Facades\Route;
 class RouteServiceProvider extends ServiceProvider
 {
     /**
-     * The controller namespace for the application.
-     *
-     * When present, controller route declarations will automatically be prefixed with this namespace.
-     *
-     * @var string|null
-     */
-    // protected $namespace = 'App\\Http\\Controllers';
-
-    /**
      * The path to the "home" route for your application.
      *
-     * This is used by Laravel authentication to redirect users after login.
+     * Typically, users are redirected here after authentication.
      *
      * @var string
      */
     public const HOME = '/';
 
     /**
-     * Define your route model bindings, pattern filters, etc.
+     * Define your route model bindings, pattern filters, and other route configuration.
      *
      * @return void
      */
@@ -38,13 +29,11 @@ class RouteServiceProvider extends ServiceProvider
         $this->configureRateLimiting();
 
         $this->routes(function () {
-            Route::prefix('api')
-                ->middleware('api')
-                ->namespace($this->namespace)
+            Route::middleware('api')
+                ->prefix('api')
                 ->group(base_path('routes/api.php'));
 
             Route::middleware('web')
-                ->namespace($this->namespace)
                 ->group(base_path('routes/web.php'));
         });
     }
@@ -57,15 +46,15 @@ class RouteServiceProvider extends ServiceProvider
     protected function configureRateLimiting()
     {
         RateLimiter::for('api', function (Request $request) {
-            return Limit::perMinute(200)->by(optional($request->user())->id ?: $request->ip());
+            return Limit::perMinute(200)->by($request->user()?->id ?: $request->ip());
         });
 
         RateLimiter::for('password_reset', function (Request $request) {
-            return Limit::perMinutes(30, 5)->by(optional($request->user())->id ?: $request->ip());
+            return Limit::perMinutes(30, 5)->by($request->user()?->id ?: $request->ip());
         });
 
         RateLimiter::for('password_email', function (Request $request) {
-            return Limit::perMinutes(30, 5)->by(optional($request->user())->id ?: $request->ip());
+            return Limit::perMinutes(30, 5)->by($request->user()?->id ?: $request->ip());
         });
     }
 }