+++ resources/js/components/Room/BrowserNotification.vue
@@ -1,8 +1,8 @@
 <template>
   <div v-if="notificationSupport && !running" class="w-100 text-center my-2">
-    <b-button v-if="!notificationEnabled" block @click="enableNotification"><i class="fas fa-bell"></i> {{ $t('rooms.notification.enable') }} </b-button>
+    <b-button v-if="!notificationEnabled" block @click="enableNotification"><i class="fa-solid fa-bell"></i> {{ $t('rooms.notification.enable') }} </b-button>
     <b-alert v-else variant="success" show>
-      <i class="fas fa-bell"></i> {{ $t('rooms.notification.enabled') }}
+      <i class="fa-solid fa-bell"></i> {{ $t('rooms.notification.enabled') }}
     </b-alert>
   </div>
 </template>