+++ resources/js/views/settings/serverPools/View.vue
@@ -17,7 +17,7 @@
             ref="reloadServerPool"
             @click="load()"
           >
-            <b-icon-arrow-clockwise></b-icon-arrow-clockwise> {{ $t('app.reload') }}
+            <i class="fa-solid fa-sync"></i> {{ $t('app.reload') }}
           </b-button>
         </div>
       </template>
@@ -74,7 +74,7 @@
                   <h5 class='d-inline mr-1 mb-1'>
                     <b-badge variant='primary'>
                       {{ option.name }}
-                      <span @click='remove(option)'><b-icon-x :aria-label="$t('settings.serverPools.removeServer')"></b-icon-x></span>
+                      <span @click='remove(option)'><i class="fa-solid fa-xmark" :aria-label="$t('settings.serverPools.removeServer')"></i></span>
                     </b-badge>
                   </h5>
                 </template>
@@ -83,13 +83,13 @@
                     :disabled='serversLoading || currentPage === 1'
                     variant='outline-secondary'
                     @click='loadServers(Math.max(1, currentPage - 1))'>
-                    <i class='fas fa-arrow-left'></i> {{ $t('app.previousPage') }}
+                    <i class='fa-solid fa-arrow-left'></i> {{ $t('app.previousPage') }}
                   </b-button>
                   <b-button
                     :disabled='serversLoading || !hasNextPage'
                     variant='outline-secondary'
                     @click='loadServers(currentPage + 1)'>
-                    <i class='fas fa-arrow-right'></i> {{ $t('app.nextPage') }}
+                    <i class='fa-solid fa-arrow-right'></i> {{ $t('app.nextPage') }}
                   </b-button>
                 </template>
               </multiselect>
@@ -98,7 +98,7 @@
                   v-if="serversLoadingError"
                   @click="loadServers(currentPage)"
                   variant="outline-secondary"
-                ><i class="fas fa-sync"></i></b-button>
+                ><i class="fa-solid fa-sync"></i></b-button>
               </b-input-group-append>
             </b-input-group>
             <template slot='invalid-feedback'><div v-html="fieldError('servers', true)"></div></template>
@@ -111,7 +111,7 @@
                 :disabled='isBusy'
                 variant='secondary'
                 @click="$router.push({ name: 'settings.server_pools' })">
-                <i class='fas fa-arrow-left'></i> {{ $t('app.back') }}
+                <i class='fa-solid fa-arrow-left'></i> {{ $t('app.back') }}
               </b-button>
               <b-button
                 :disabled='isBusy || modelLoadingError || serversLoadingError || serversLoading'
@@ -119,7 +119,7 @@
                 type='submit'
                 class='ml-1'
                 v-if='!viewOnly'>
-                <i class='fas fa-save'></i> {{ $t('app.save') }}
+                <i class='fa-solid fa-save'></i> {{ $t('app.save') }}
               </b-button>
             </b-col>
           </b-row>