+++ CHANGELOG.md
@@ -5,6 +5,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Changed
+- Upgrade to Laravel 9 ([#226],[#227])
+- Breaking: Drop support for PHP 7.4 and PHP 8.0 ([#226],[#227])
+- Breaking: Replace laravel homestead with laravel sail ([#225],[#228])
+- Remove git hooks ([#217],[#228])
+
+## [1.9.4] - 2022-06-30
+### Changed
+- Update composer dependencies ([#221],[#222])
+- Update php-cs-fixer ([#221],[#222])
+- Add php 8.1 support ([#221],[#222])
+- Unify used icons ([#223],[#224])
+- Update node dependencies and fontawesome ([#223],[#224])
+
 ## [1.9.3] - 2022-04-14
 ### Fixed
 - Server error 500 on login, if entryUUID in ldap changed ([#212], [#213])
@@ -267,8 +282,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#205]: https://github.com/THM-Health/PILOS/pull/205
 [#212]: https://github.com/THM-Health/PILOS/issues/212
 [#213]: https://github.com/THM-Health/PILOS/pull/213
+[#217]: https://github.com/THM-Health/PILOS/issues/217
+[#221]: https://github.com/THM-Health/PILOS/issues/221
+[#222]: https://github.com/THM-Health/PILOS/pull/222
+[#223]: https://github.com/THM-Health/PILOS/issues/223
+[#224]: https://github.com/THM-Health/PILOS/pull/224
+[#225]: https://github.com/THM-Health/PILOS/issues/225
+[#226]: https://github.com/THM-Health/PILOS/issues/226
+[#227]: https://github.com/THM-Health/PILOS/pull/227
+[#228]: https://github.com/THM-Health/PILOS/pull/228
 
-[unreleased]: https://github.com/THM-Health/PILOS/compare/v1.9.3...HEAD
+[unreleased]: https://github.com/THM-Health/PILOS/compare/v1.9.4...HEAD
 [1.0.0]: https://github.com/THM-Health/PILOS/releases/tag/v1.0.0
 [1.1.0]: https://github.com/THM-Health/PILOS/releases/tag/v1.1.0
 [1.2.0]: https://github.com/THM-Health/PILOS/releases/tag/v1.2.0
@@ -284,3 +308,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.9.1]: https://github.com/THM-Health/PILOS/releases/tag/v1.9.1
 [1.9.2]: https://github.com/THM-Health/PILOS/releases/tag/v1.9.2
 [1.9.3]: https://github.com/THM-Health/PILOS/releases/tag/v1.9.3
+[1.9.4]: https://github.com/THM-Health/PILOS/releases/tag/v1.9.4